backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client>=1.7.1
openai==1.12.0
python-dotenv==1.0.0
tiktoken>=0.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.26.0
psycopg2-binary>=2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]>=2.0.0

